% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeTemperatureData.R
\name{analyzeTemperatureData}
\alias{analyzeTemperatureData}
\title{Analyze Temperature Data}
\usage{
analyzeTemperatureData(country)
}
\arguments{
\item{country}{The name of the country to analyze.}
}
\value{
A plot of the temperature data.
}
\description{
This function analyzes temperature data for a given country.
}
\examples{
analyzeTemperatureData("Japan")
}
